lambda-backend/requirements.txt
certifi==2025.6.15
charset-normalizer==3.4.2
idna==3.10
paho-mqtt==2.1.0
requests==2.32.4
urllib3==2.5.0
